jpa tomcat

+++ src/main/java/com/proxybanque_KO_JFA/views/ViewClientsServlet.java
@@ -52,7 +52,7 @@ public class ViewClientsServlet extends HttpServlet {
 		HttpSession session = request.getSession(false);
 
 		if (session != null && session.getAttribute("user") != null) {
-			int idConseiller = (Integer) session.getAttribute("idCons");
+			Long idConseiller = (Long) session.getAttribute("idCons");
 
 			if (idConseiller != 0) {
 				try {
